increaseCount 와 View Layer 분리
