+++ src/components/Page.jsx
@@ -8,8 +8,7 @@ export default function Page({
 }) {
   return (
     <div>
-      <p>Hello, world!</p>
-      <p>Hi!</p>
+      <p>Counter</p>
       <Counter
         count={count}
         onClick={onClick}
